requirements.txt
# Requirements mainly for Django, DRF, and pydantic.
# Other packages required mainly for testing or documentation.

Django>=3.2.0
djangorestframework>=3.11.0
pydantic>=1.8
typing_extensions==4.0.0
pytest>=6.2.0
coverage==6.2
sphinx==4.3.1
sphinx-rtd-theme==1.0.0